refactor(chat): message-pipeline 600줄 가드 복구 — 헬퍼 2개 모듈 분리
외부 모델 모드 라우팅 추가로 620줄 초과 → mode-external-client(외부 모델
해석)·user-context-blocks(memory/CI 조립) 추출해 538줄로 복구. 동작 무변경.
